Split multiagent handler files
